refactor(calendar): extract event form, navigation, and debug logic into custom hooks
Moved event form handling, calendar navigation, and debug tools logic into separate custom hooks (useEventForm, useCalendarNavigation, and setupDebugTools) to improve code maintainability and reusability. Updated the CalendarMain component to utilize these hooks, reducing its complexity and improving readability. Also added documentation updates to reflect the new hooks.
