+++ docs/dev/stages/stage-4.md
@@ -130,6 +130,8 @@ atlas-core/
 │   │   ├── use-calendar-events.jsx  # Hook para eventos del calendario
 │   │   ├── use-time-grid.jsx        # Hook para rejilla temporal
 │   │   ├── use-event-drag.jsx       # Hook para arrastrar eventos
+│   │   ├── use-event-form.jsx    # Hook para formulario de eventos
+│   │   ├── use-calendar-navigation.jsx # Hook para navegación en el calendario
 │   │   ├── use-event-resize.jsx     # Hook para redimensionar eventos
 │   │   ├── use-theme.jsx            # Hook para gestión de temas
 │   │   └── use-ui-utils.jsx         # Hook para utilidades UI
@@ -137,6 +139,7 @@ atlas-core/
 │   ├── utils/                       # Utilidades
 │   │   ├── date-utils.js            # Utilidades de fechas
 │   │   ├── time-utils.js            # Utilidades de tiempo
+│   │   ├── debug-utils.js        # Utilidades de depuración
 │   │   ├── event-utils.js           # Utilidades para eventos
 │   │   ├── theme-utils.js           # Utilidades para temas
 │   │   ├── storage-utils.js         # Utilidades de almacenamiento
