+++ src/components/calendar/calendar-main.jsx
@@ -6,6 +6,9 @@ import DayView from './day-view';
 import EventForm from './event-form';
 import SnapControl from './snap-control';
 import useCalendarEvents from '../../hooks/use-calendar-events';
+import useCalendarNavigation from '../../hooks/use-calendar-navigation';
+import useEventForm from '../../hooks/use-event-form';
+import { setupDebugTools } from '../../utils/debug-utils';
 import '../../styles/calendar/calendar-main.css';
 
 /**
@@ -22,21 +25,36 @@ function CalendarMain() {
     saveEvents 
   } = useCalendarEvents();
 
-  // Estados principales
-  const [currentDate, setCurrentDate] = useState(new Date());
-  const [selectedEvent, setSelectedEvent] = useState(null);
-  const [showEventForm, setShowEventForm] = useState(false);
-  const [formError, setFormError] = useState('');
+  // Estado de la vista y snap
   const [view, setView] = useState('week'); // 'week' o 'day'
-  const [selectedDay, setSelectedDay] = useState(new Date());
   const [snapValue, setSnapValue] = useState(0); // 0 = desactivado por defecto
-  const [newEvent, setNewEvent] = useState({
-    id: '',
-    title: '',
-    start: '',
-    end: '',
-    color: '#2d4b94' // Color predeterminado (Azul Atlas)
-  });
+  
+  // Usar hook de navegación
+  const {
+    currentDate,
+    selectedDay,
+    setSelectedDay,
+    goToPreviousWeek,
+    goToNextWeek,
+    goToCurrentWeek,
+    goToPreviousDay,
+    goToNextDay,
+    goToToday
+  } = useCalendarNavigation();
+  
+  // Usar hook de formulario de eventos
+  const {
+    selectedEvent,
+    showEventForm,
+    formError,
+    newEvent,
+    handleEventClick,
+    handleCellClick,
+    handleEventFormChange,
+    handleCloseForm,
+    handleSaveEvent,
+    handleDeleteEvent
+  } = useEventForm(createEvent, updateEvent, deleteEvent);
 
   // Registrar módulo al iniciar
   useEffect(() => {
@@ -54,65 +72,10 @@ function CalendarMain() {
     };
   }, []);
 
-  // Exponer funciones de depuración
+  // Configurar herramientas de depuración
   useEffect(() => {
-    if (typeof window !== 'undefined') {
-      window.debugAtlas = {
-        getEvents: () => events,
-        createTestEvent: () => {
-          const now = new Date();
-          const start = new Date(now);
-          start.setHours(start.getHours() + 1, 0, 0, 0);
-          
-          const end = new Date(start);
-          end.setHours(end.getHours() + 1);
-          
-          const testEvent = {
-            title: `Evento de prueba ${Date.now()}`,
-            start: start.toISOString(),
-            end: end.toISOString(),
-            color: '#FF5722'
-          };
-          
-          const createdEvent = createEvent(testEvent);
-          console.log('Evento de prueba creado:', createdEvent);
-          return createdEvent;
-        },
-        forceUpdate: (eventId, hoursToMove) => {
-          const eventToUpdate = events.find(e => e.id === eventId);
-          if (!eventToUpdate) {
-            console.error('Evento no encontrado:', eventId);
-            return null;
-          }
-          
-          const startDate = new Date(eventToUpdate.start);
-          const endDate = new Date(eventToUpdate.end);
-          
-          startDate.setHours(startDate.getHours() + hoursToMove);
-          endDate.setHours(endDate.getHours() + hoursToMove);
-          
-          const updatedEvent = {
-            ...eventToUpdate,
-            start: startDate.toISOString(),
-            end: endDate.toISOString()
-          };
-          
-          const result = updateEvent(eventId, updatedEvent);
-          console.log('Actualización forzada:', result);
-          return result;
-        },
-        saveAllEvents: () => {
-          saveEvents(events);
-          console.log('Eventos guardados manualmente');
-        }
-      };
-    }
-    
-    return () => {
-      if (typeof window !== 'undefined') {
-        delete window.debugAtlas;
-      }
-    };
+    const cleanup = setupDebugTools(events, createEvent, updateEvent, saveEvents);
+    return cleanup;
   }, [events]);
 
   // Cambiar entre vistas
@@ -120,217 +83,6 @@ function CalendarMain() {
     setView(newView);
     if (date) {
       setSelectedDay(new Date(date));
-    }
-  };
-
-  // Navegación entre semanas
-  const goToPreviousWeek = () => {
-    const newDate = new Date(currentDate);
-    newDate.setDate(newDate.getDate() - 7);
-    setCurrentDate(newDate);
-  };
-
-  const goToNextWeek = () => {
-    const newDate = new Date(currentDate);
-    newDate.setDate(newDate.getDate() + 7);
-    setCurrentDate(newDate);
-  };
-
-  const goToCurrentWeek = () => {
-    setCurrentDate(new Date());
-  };
-
-  // Manejar navegación de días en vista diaria
-  const goToPreviousDay = () => {
-    const prevDay = new Date(selectedDay);
-    prevDay.setDate(prevDay.getDate() - 1);
-    setSelectedDay(prevDay);
-  };
-
-  const goToNextDay = () => {
-    const nextDay = new Date(selectedDay);
-    nextDay.setDate(nextDay.getDate() + 1);
-    setSelectedDay(nextDay);
-  };
-
-  const goToToday = () => {
-    setSelectedDay(new Date());
-  };
-
-  // Manejar clic en celda para crear evento
-  const handleCellClick = (day, hour) => {
-    try {
-      const startDate = new Date(day);
-      startDate.setHours(hour, 0, 0, 0);
-      
-      const endDate = new Date(startDate);
-      endDate.setHours(hour + 1, 0, 0, 0);
-      
-      // Formatear fechas para inputs
-      const formatDateForInput = (date) => {
-        const year = date.getFullYear();
-        const month = String(date.getMonth() + 1).padStart(2, '0');
-        const day = String(date.getDate()).padStart(2, '0');
-        const hours = String(date.getHours()).padStart(2, '0');
-        const minutes = String(date.getMinutes()).padStart(2, '0');
-        
-        return `${year}-${month}-${day}T${hours}:${minutes}`;
-      };
-      
-      setSelectedEvent(null);
-      setFormError('');
-      
-      setNewEvent({
-        id: '',
-        title: 'Nuevo evento',
-        start: startDate.toISOString(),
-        end: endDate.toISOString(),
-        startFormatted: formatDateForInput(startDate),
-        endFormatted: formatDateForInput(endDate),
-        color: '#2d4b94'
-      });
-      
-      setShowEventForm(true);
-    } catch (error) {
-      console.error('Error al manejar clic en celda:', error);
-    }
-  };
-
-  // Manejar clic en evento existente
-  const handleEventClick = (event) => {
-    try {
-      const startDate = new Date(event.start);
-      const endDate = new Date(event.end);
-      
-      if (isNaN(startDate.getTime()) || isNaN(endDate.getTime())) {
-        console.error('Fechas de evento inválidas:', event);
-        return;
-      }
-      
-      // Formatear fechas para inputs
-      const formatDateForInput = (date) => {
-        const year = date.getFullYear();
-        const month = String(date.getMonth() + 1).padStart(2, '0');
-        const day = String(date.getDate()).padStart(2, '0');
-        const hours = String(date.getHours()).padStart(2, '0');
-        const minutes = String(date.getMinutes()).padStart(2, '0');
-        
-        return `${year}-${month}-${day}T${hours}:${minutes}`;
-      };
-      
-      setSelectedEvent({...event});
-      setFormError('');
-      
-      setNewEvent({
-        id: event.id,
-        title: event.title || '',
-        start: event.start,
-        end: event.end,
-        color: event.color || '#2d4b94',
-        startFormatted: formatDateForInput(startDate),
-        endFormatted: formatDateForInput(endDate)
-      });
-      
-      setShowEventForm(true);
-    } catch (error) {
-      console.error('Error al manejar clic en evento:', error, event);
-    }
-  };
-
-  // Actualizar datos del formulario
-  const handleEventFormChange = (e) => {
-    try {
-      const { name, value } = e.target;
-      
-      if (name === 'start' || name === 'end') {
-        const date = new Date(value);
-        
-        if (isNaN(date.getTime())) {
-          console.error('Fecha inválida:', value);
-          return;
-        }
-        
-        setNewEvent(prev => ({ 
-          ...prev, 
-          [name]: date.toISOString(),
-          [`${name}Formatted`]: value 
-        }));
-      } else {
-        setNewEvent(prev => ({ ...prev, [name]: value }));
-      }
-    } catch (error) {
-      console.error('Error al manejar cambio en formulario:', error);
-    }
-  };
-
-  // Cerrar y reiniciar formulario
-  const handleCloseForm = () => {
-    setShowEventForm(false);
-    setSelectedEvent(null);
-    setFormError('');
-    setNewEvent({
-      id: '',
-      title: '',
-      start: '',
-      end: '',
-      startFormatted: '',
-      endFormatted: '',
-      color: '#2d4b94'
-    });
-  };
-
-  // Guardar evento
-  const handleSaveEvent = () => {
-    try {
-      setFormError('');
-      
-      // Validar título
-      if (!newEvent.title.trim()) {
-        setFormError('El título del evento no puede estar vacío');
-        return;
-      }
-      
-      // Validar fechas
-      const startDate = new Date(newEvent.start);
-      const endDate = new Date(newEvent.end);
-      
-      if (endDate < startDate) {
-        setFormError('La hora de fin no puede ser anterior a la hora de inicio');
-        return;
-      }
-      
-      const eventToSave = {
-        id: newEvent.id || Date.now().toString(),
-        title: newEvent.title.trim(),
-        start: newEvent.start,
-        end: newEvent.end,
-        color: newEvent.color || '#2d4b94'
-      };
-      
-      if (selectedEvent && selectedEvent.id) {
-        // Actualizar existente
-        updateEvent(selectedEvent.id, eventToSave);
-      } else {
-        // Crear nuevo
-        createEvent(eventToSave);
-      }
-      
-      handleCloseForm();
-    } catch (error) {
-      console.error('Error al guardar evento:', error);
-      setFormError('Ocurrió un error al guardar el evento');
-    }
-  };
-
-  // Eliminar evento
-  const handleDeleteEvent = () => {
-    try {
-      if (selectedEvent) {
-        deleteEvent(selectedEvent.id);
-        handleCloseForm();
-      }
-    } catch (error) {
-      console.error('Error al eliminar evento:', error);
     }
   };
