+++ src/hooks/use-event-form.jsx
@@ -1,0 +1,192 @@
+import { useState } from 'react';
+import { formatDateForInput } from '../utils/date-utils';
+
+/**
+ * Hook personalizado para gestionar el formulario de eventos
+ * @param {Function} createEvent - Función para crear eventos
+ * @param {Function} updateEvent - Función para actualizar eventos
+ * @param {Function} deleteEvent - Función para eliminar eventos
+ * @returns {Object} - Funciones y estados para el formulario
+ */
+function useEventForm(createEvent, updateEvent, deleteEvent) {
+  const [selectedEvent, setSelectedEvent] = useState(null);
+  const [showEventForm, setShowEventForm] = useState(false);
+  const [formError, setFormError] = useState('');
+  const [newEvent, setNewEvent] = useState({
+    id: '',
+    title: '',
+    start: '',
+    end: '',
+    color: '#2d4b94' // Color predeterminado (Azul Atlas)
+  });
+  
+  // Manejar clic en celda para crear evento
+  const handleCellClick = (day, hour) => {
+    try {
+      const startDate = new Date(day);
+      startDate.setHours(hour, 0, 0, 0);
+      
+      const endDate = new Date(startDate);
+      endDate.setHours(hour + 1, 0, 0, 0);
+      
+      setSelectedEvent(null);
+      setFormError('');
+      
+      setNewEvent({
+        id: '',
+        title: 'Nuevo evento',
+        start: startDate.toISOString(),
+        end: endDate.toISOString(),
+        startFormatted: formatDateForInput(startDate),
+        endFormatted: formatDateForInput(endDate),
+        color: '#2d4b94'
+      });
+      
+      setShowEventForm(true);
+    } catch (error) {
+      console.error('Error al manejar clic en celda:', error);
+    }
+  };
+
+  // Manejar clic en evento existente
+  const handleEventClick = (event) => {
+    try {
+      const startDate = new Date(event.start);
+      const endDate = new Date(event.end);
+      
+      if (isNaN(startDate.getTime()) || isNaN(endDate.getTime())) {
+        console.error('Fechas de evento inválidas:', event);
+        return;
+      }
+      
+      setSelectedEvent({...event});
+      setFormError('');
+      
+      setNewEvent({
+        id: event.id,
+        title: event.title || '',
+        start: event.start,
+        end: event.end,
+        color: event.color || '#2d4b94',
+        startFormatted: formatDateForInput(startDate),
+        endFormatted: formatDateForInput(endDate)
+      });
+      
+      setShowEventForm(true);
+    } catch (error) {
+      console.error('Error al manejar clic en evento:', error, event);
+    }
+  };
+
+  // Actualizar datos del formulario
+  const handleEventFormChange = (e) => {
+    try {
+      const { name, value } = e.target;
+      
+      if (name === 'start' || name === 'end') {
+        const date = new Date(value);
+        
+        if (isNaN(date.getTime())) {
+          console.error('Fecha inválida:', value);
+          return;
+        }
+        
+        setNewEvent(prev => ({ 
+          ...prev, 
+          [name]: date.toISOString(),
+          [`${name}Formatted`]: value 
+        }));
+      } else {
+        setNewEvent(prev => ({ ...prev, [name]: value }));
+      }
+    } catch (error) {
+      console.error('Error al manejar cambio en formulario:', error);
+    }
+  };
+
+  // Cerrar y reiniciar formulario
+  const handleCloseForm = () => {
+    setShowEventForm(false);
+    setSelectedEvent(null);
+    setFormError('');
+    setNewEvent({
+      id: '',
+      title: '',
+      start: '',
+      end: '',
+      startFormatted: '',
+      endFormatted: '',
+      color: '#2d4b94'
+    });
+  };
+
+  // Guardar evento
+  const handleSaveEvent = () => {
+    try {
+      setFormError('');
+      
+      // Validar título
+      if (!newEvent.title.trim()) {
+        setFormError('El título del evento no puede estar vacío');
+        return;
+      }
+      
+      // Validar fechas
+      const startDate = new Date(newEvent.start);
+      const endDate = new Date(newEvent.end);
+      
+      if (endDate < startDate) {
+        setFormError('La hora de fin no puede ser anterior a la hora de inicio');
+        return;
+      }
+      
+      const eventToSave = {
+        id: newEvent.id || Date.now().toString(),
+        title: newEvent.title.trim(),
+        start: newEvent.start,
+        end: newEvent.end,
+        color: newEvent.color || '#2d4b94'
+      };
+      
+      if (selectedEvent && selectedEvent.id) {
+        // Actualizar existente
+        updateEvent(selectedEvent.id, eventToSave);
+      } else {
+        // Crear nuevo
+        createEvent(eventToSave);
+      }
+      
+      handleCloseForm();
+    } catch (error) {
+      console.error('Error al guardar evento:', error);
+      setFormError('Ocurrió un error al guardar el evento');
+    }
+  };
+
+  // Eliminar evento
+  const handleDeleteEvent = () => {
+    try {
+      if (selectedEvent) {
+        deleteEvent(selectedEvent.id);
+        handleCloseForm();
+      }
+    } catch (error) {
+      console.error('Error al eliminar evento:', error);
+    }
+  };
+  
+  return {
+    selectedEvent,
+    showEventForm,
+    formError,
+    newEvent,
+    handleEventClick,
+    handleCellClick,
+    handleEventFormChange,
+    handleCloseForm,
+    handleSaveEvent,
+    handleDeleteEvent
+  };
+}
+
+export default useEventForm;
