+++ src/utils/date-utils.js
@@ -124,3 +124,19 @@ export function isSameDay(date1, date2) {
     d1.getDate() === d2.getDate()
   );
 }
+
+// Añadir esta función al final del archivo
+/**
+ * Formatea una fecha para input datetime-local
+ * @param {Date} date - Fecha a formatear
+ * @returns {string} - Fecha formateada en formato "YYYY-MM-DDTHH:MM"
+ */
+export function formatDateForInput(date) {
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, '0');
+  const day = String(date.getDate()).padStart(2, '0');
+  const hours = String(date.getHours()).padStart(2, '0');
+  const minutes = String(date.getMinutes()).padStart(2, '0');
+  
+  return `${year}-${month}-${day}T${hours}:${minutes}`;
+}
